Gonna do a combat rework of sorts:
All units will attack the one they initially target, regardless of if it is already killed by the time an attacker's damage is calculated. Death will be emitted as an "extra mutation", a system I'll implement to allow mutations to themselves create additional mutations that will be fed back into the Stepper.

Along with this change, we'll make it so that units attack the other if and only if they are moving into its space. We will overhaul the Defender tag to mean "performs combat against attackers, but cannot attack."